api/requirements.txt
# cache-bust-2026-04-27
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openpyxl>=3.1.2
pydantic>=2.6.0
scikit-learn>=1.4.0
lightgbm>=4.0.0
xgboost>=2.0.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0
scipy>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
httpx>=0.27.0
soccerdata>=1.8
colorama==0.4.6
